fix details
